worker/requirements.txt
kafka-python
nltk
mysql-connector-python